Move makePreviousOsIdTooltipStyles into styles.js
